pyproject.toml
[project]
name = "ai-call-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
    "twilio>=9.8.0",
]
